Created basic structure for landing page
HTML and CSS file setup. The basic structure of the landing page
has been created, four main divs for the four color coded sections
of the page, with further divs nested inside to differentiate the sections.

Flexbox is being used to arrange the various sections. At the moment I feel
like I may be using one too many div elements to organize the structure of the
page but so far there are no issues.